role for appointment added

+++ backend/controller/appointment/appointmentsController.js
@@ -1,7 +1,14 @@
 const sqlite3 = require('sqlite3').verbose();
 const db = new sqlite3.Database('demo.db');
+const {getGrants} = require('../../utitlity/roleManager');
 
 exports.saveNewAppointment = async (req, res) => {
+
+  const permission = getGrants.can(req.user.role).createAny("appointment");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden");
+  }
+
   const appointmentData = {
     patientId: req.body.patientId,
     appointmentTypeId: req.body.appointmentTypeId,
@@ -27,6 +34,12 @@ exports.saveNewAppointment = async (req, res) => {
 };
 
 exports.updateAppointment = async (req, res) => {
+  const permission = getGrants.can(req.user.role).updateAny("appointment");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden");
+  }
+
+
   const appointmentData = {
     patientId: req.body.patientId,
     appointmentTypeId: req.body.appointmentTypeId,
@@ -51,6 +64,12 @@ exports.updateAppointment = async (req, res) => {
 };
 
 exports.deleteAppointmentById = async (req, res) => {
+  const permission = getGrants.can(req.user.role).deleteAny("appointment");
+  if (!permission.granted) {
+    return res.status(403).json("Access forbidden");
+  }
+
+
   await db.run("DELETE from Appointment WHERE id= ?", req.params['appointmentId'], function (err) {
     if (err) {
       res.json(err.message).status(404);
@@ -60,26 +79,34 @@ exports.deleteAppointmentById = async (req, res) => {
   });
 };
 
+function hasReadAnyAccess() {
+  return getGrants.can(req.user.role).readAny("appointment");
+}
+
+
 exports.getPatientAppointment = async (req, res) => {
+  if (!hasReadAnyAccess()) {
+    return res.status(403).json("Access forbidden");
+  }
+
   await db.all("select * from appointment where patientId = ?", [req.params['patientId']], function (err, rows) {
     return res.json(rows);
   });
 };
 
 exports.getDoctorAppointments = async (req, res) => {
+
   await db.all("select * from appointment where servedBy = ?", [req.params['doctorId']], function (err, rows) {
     return res.json(rows);
   });
 };
 
 exports.getAllAppointments = async (req, res) => {
-  await db.all("SELECT * FROM appointment", function (err, rows) {
-    return res.json(rows);
-  });
-};
+  if (!hasReadAnyAccess()) {
+    return res.status(403).json("Access forbidden");
+  }
 
-exports.getAppointmentStatus = async (req, res) => {
-  await db.all("SELECT * FROM appointmentStatus", function (err, rows) {
+  await db.all("SELECT * FROM appointment", function (err, rows) {
     return res.json(rows);
   });
 };
